// src/app/interface/datos.interface.ts
export interface Datos {
  uid?: string;
  tiempo: number;
  nombre: string;
  fecha: string;
  lugar: string;
  cocacola: number;
  sprite: number;
  fanta: number;
  ciel: number;
  sidral: number;
  delvalle: number;
  fuze: number;
  powerade: number;
  glaceau: number;
  limon: number;
  naranja: number;
  santaclara: number;
  fresca: number;
  ades: number;
}